+++ netology_5.racegame/Camel.h
@@ -5,6 +5,6 @@ class Camel :
 {
 public:
     Camel();
-    int getTotalRaceTime() override;
+    double getTotalRaceTime() override;
 };
