+++ netology_5.racegame/Transport.h
@@ -3,19 +3,19 @@ class Transport
 {
 protected:
     // Общие данные
-    int totalRestingTime = 0;
-    int restCount = 0;
+    double totalRestingTime = 0;
+    double restCount = 0;
     int distance = 0;
-    int raceTimeWoRest = 0;
-    int totalRaceTime = 0;
+    double raceTimeWoRest = 0;
+    double totalRaceTime = 0;
     // Данные для ТС
     int speed = 0;
-    int timeToRest = 0;
-    int resting1stTime = 0;
-    int resting2ndTime = 0;
-    int restingOtherTime = 0;
+    double timeToRest = 0;
+    double resting1stTime = 0;
+    double resting2ndTime = 0;
+    double restingOtherTime = 0;
 public:
     bool setDistance(int distance);
-    virtual int getTotalRaceTime();
+    virtual double getTotalRaceTime();
 };
